+++ src/scripts/plus.js
@@ -4,21 +4,12 @@ import autoInit from './config/auto-init';
 /**
  * Components
  */
-import UiAutocomplete from './components/autocomplete/autocomplete';
 import UiSkeleton from './components/skeleton/skeleton';
-import UiDatepicker from './components/datepicker/datepicker';
-import UiRangepicker from './components/rangepicker/rangepicker';
 import UiAlert from './components/alert/alert';
 import UiCollapse from './components/collapse/collapse';
 import UiEditor from './components/editor/editor';
 import UiTree from './components/tree/tree';
 import UiTreeNode from './components/tree/tree-node';
-/**
- * Plugins
- */
-import $alert, { useAlert } from './plugins/alert';
-import $confirm, { useConfirm } from './plugins/confirm';
-import $toast, { useToast } from './plugins/toast';
 /**
  * Directives
  */
@@ -29,21 +20,14 @@ import vDebounce from './directives/debounce';
 
 const Components = {
   UiAlert,
-  UiAutocomplete,
   UiCollapse,
-  UiDatepicker,
   UiEditor,
-  UiRangepicker,
   UiSkeleton,
   UiTree,
   UiTreeNode
 };
 
-const Plugins = {
-  $alert,
-  $confirm,
-  $toast
-};
+const Plugins = {};
 
 const Directives = {
   vAnchor,
@@ -68,17 +52,5 @@ autoInit(BalmUIPlus);
 
 export default BalmUIPlus;
 export { version, install };
-export {
-  UiAlert,
-  UiAutocomplete,
-  UiCollapse,
-  UiDatepicker,
-  UiEditor,
-  UiRangepicker,
-  UiSkeleton,
-  UiTree,
-  UiTreeNode
-};
-export { $alert, $confirm, $toast };
-export { useAlert, useConfirm, useToast };
+export { UiAlert, UiCollapse, UiEditor, UiSkeleton, UiTree, UiTreeNode };
 export { vAnchor, vCopy, vDebounce, vLongpress };